commented validateInclusionAuxData & computeExpectedAuxDataWithDeal
Okay, in the current flow, these required conditions (which are giving compile time error) are in the `validateInclusionAuxData` function, which is an internal function of `computeExpectedAuxDataWithDeal`. The fun fact is that `computeExpectedAuxDataWithDeal` is commented out in the `EdgeAggregatorOracle.sol` contract. Therefore, it doesn't affect the flow. However, I do suggest either deleting or commenting out these functions in the `proof.sol` contract to remove confusion and compile-time errors.